requirements.txt
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
cycler==0.10.0
fire==0.3.1
idna==2.10
jieba==0.42.1
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.1
Pillow==7.2.0
prettytable==0.7.2
psutil==5.7.2
py-cpuinfo==7.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.13.1
python-iotbot==2.1.0
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.24.0
retrying==1.3.3
schedule==0.6.0
six==1.15.0
termcolor==1.1.0
urllib3==1.25.9
websocket-client==0.57.0
wordcloud==1.8.0
torchnet==0.0.4
fire==0.1.3
torch==1.0.1